+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha26] - 2019-09-20
+### Change
+ - Public ID is required during image upload.
+
+### Add
+ - `StringHelper` class for string functions.
+ - `UploadHelper` class to provide image public ID generation opportunity.
+
 ## [1.0.0-alpha25] - 2019-08-21
 ### Fix
  - URL encode fixes